requirements.txt
uvicorn
openai
sentence-transformers
faiss-cpu
pypandoc
python-dotenv
jinja2
markdown
